+++ addon-test-support/helpers.js
@@ -5,7 +5,8 @@ import {
   selectChoose as _selectChoose,
   selectSearch as _selectSearch,
   removeMultipleOption as _removeMultipleOption,
-  clearSelected as _clearSelected
+  clearSelected as _clearSelected,
+  getDropdownItems as _getDropdownItems
 } from './index';
 /**
  * @private
@@ -70,6 +71,18 @@ export async function touchTrigger() {
 
 export async function selectChoose(cssPathOrTrigger, valueOrSelector, optionIndex) {
   return _selectChoose(cssPathOrTrigger, valueOrSelector, optionIndex);
+}
+
+/* *
+ * @param {String} selector CSS3 selector of the elements to check the content
+ * @returns {Array} returns all the elements present in the dropdown
+ *
+ * Usage in tests:
+ * let options = await getDropdownItems('.ember-power-select-trigger');
+ * assert.deepEqual(options, [1, 2, 3]);
+ * */
+export async function getDropdownItems(cssPathOrTrigger) {
+  return _getDropdownItems(cssPathOrTrigger);
 }
 
 export async function selectSearch(cssPathOrTrigger, value) {