+++ addon/components/power-select.ts
@@ -20,7 +20,7 @@ import {
 import { restartableTask } from 'ember-concurrency-decorators';
 // @ts-ignore
 import { timeout } from 'ember-concurrency';
-import { Dropdown, DropdownActions } from 'ember-basic-dropdown/addon/components/basic-dropdown'
+import { Dropdown, DropdownActions } from 'ember-basic-dropdown/addon/components/basic-dropdown';
 
 interface SelectActions extends DropdownActions {
   search: (term: string) => void
@@ -88,8 +88,12 @@ const isArrayable = <T>(coll: any): coll is Arrayable<T> => {
   return typeof coll.toArray === 'function';
 }
 
-const isPromiseProxy = <T>(thing: any): thing is PromiseProxy<T> => {
+const isPromiseLike = <T>(thing: any): thing is Promise<T> => {
   return typeof thing.then === 'function';
+}
+
+const isPromiseProxyLike = <T>(thing: any): thing is PromiseProxy<T> => {
+  return isPromiseLike(thing) && Object.hasOwnProperty.call(thing, 'content');
 }
 
 const isCancellablePromise = <T>(thing: any): thing is CancellablePromise<T> => {
@@ -127,6 +131,16 @@ export default class PowerSelect extends Component<PowerSelectArgs> {
   constructor(owner: unknown, args: PowerSelectArgs) {
     super(owner, args);
     assert('<PowerSelect> requires an `@onChange` function', this.args.onChange && typeof this.args.onChange === 'function');
+  }
+
+  willDestroy() {
+    if (this._lastSelectedPromise && isPromiseProxyLike(this._lastSelectedPromise)) {
+      try {
+        removeObserver(this._lastSelectedPromise, 'content', this, this._selectedObserverCallback);
+      } catch {}
+      this._lastSelectedPromise = undefined;
+    }
+    super.willDestroy.apply(this, arguments);
   }
 
   // Getters
@@ -268,7 +282,7 @@ export default class PowerSelect extends Component<PowerSelectArgs> {
   @action
   handleFocus(event: FocusEvent): void {
     if (!this.isDestroying) {
-      this.isActive = true;
+      scheduleOnce('actions', this, this._updateIsActive, true);
     }
     if (this.args.onFocus) {
       this.args.onFocus(this.storedAPI, event);
@@ -278,7 +292,7 @@ export default class PowerSelect extends Component<PowerSelectArgs> {
   @action
   handleBlur(event: FocusEvent): void {
     if (!this.isDestroying) {
-      this.isActive = false;
+      scheduleOnce('actions', this, this._updateIsActive, false);
     }
     if (this.args.onBlur) {
       this.args.onBlur(this.storedAPI, event);
@@ -299,7 +313,7 @@ export default class PowerSelect extends Component<PowerSelectArgs> {
   @action
   _updateOptions(): void {
     if (!this.args.options) return
-    if (isPromiseProxy(this.args.options)) {
+    if (isPromiseLike(this.args.options)) {
       if (this._lastOptionsPromise === this.args.options) return; // promise is still the same
       let currentOptionsPromise = this.args.options;
       this._lastOptionsPromise = currentOptionsPromise as PromiseProxy<any[]>;
@@ -332,13 +346,18 @@ export default class PowerSelect extends Component<PowerSelectArgs> {
     if (!this.args.selected) return;
     if (typeof this.args.selected.then === 'function') {
       if (this._lastSelectedPromise === this.args.selected) return; // promise is still the same
-      let currentSelectedPromise: PromiseProxy<any> = this.args.selected;
-      if (Object.hasOwnProperty.call(currentSelectedPromise, 'content')) { // seems a PromiseProxy
-        if (this._lastSelectedPromise) {
-          removeObserver(this._lastSelectedPromise, 'content', this, this._selectedObserverCallback);
-        }
-        addObserver(currentSelectedPromise, 'content', this, this._selectedObserverCallback);
+      if (this._lastSelectedPromise && isPromiseProxyLike(this._lastSelectedPromise)) {
+        removeObserver(this._lastSelectedPromise, 'content', this, this._selectedObserverCallback);
       }
+
+      let currentSelectedPromise: PromiseProxy<any> = this.args.selected;
+      currentSelectedPromise.then(() => {
+        if (this.isDestroyed || this.isDestroying) return;
+        if (isPromiseProxyLike(currentSelectedPromise)) {
+          addObserver(currentSelectedPromise, 'content', this, this._selectedObserverCallback);
+        }
+      });
+
       this._lastSelectedPromise = currentSelectedPromise;
       this._lastSelectedPromise.then(resolvedSelected => {
         if (this._lastSelectedPromise === currentSelectedPromise) {
@@ -348,7 +367,10 @@ export default class PowerSelect extends Component<PowerSelectArgs> {
       });
     } else {
       this._resolvedSelected = undefined;
-      this._highlight(this.args.selected)
+      // Don't highlight args.selected array on multi-select
+      if (!Array.isArray(this.args.selected)) {
+        this._highlight(this.args.selected);
+      }
     }
   }
 
@@ -435,7 +457,7 @@ export default class PowerSelect extends Component<PowerSelectArgs> {
       return;
     }
     let searchResult = this.args.search(term, this.storedAPI);
-    if (searchResult && isPromiseProxy(searchResult)) {
+    if (searchResult && isPromiseLike(searchResult)) {
       this.loading = true;
       if (this._lastSearchPromise !== undefined && isCancellablePromise(this._lastSearchPromise)) {
         this._lastSearchPromise.cancel(); // Cancel ember-concurrency tasks
@@ -457,6 +479,7 @@ export default class PowerSelect extends Component<PowerSelectArgs> {
     } else {
       this.lastSearchedText = term;
       this._searchResult = searchResult;
+      this._resetHighlighted();
     }
   }
 
@@ -532,6 +555,9 @@ export default class PowerSelect extends Component<PowerSelectArgs> {
     return filterOptions(options || [], term, optionMatcher, skipDisabled);
   }
 
+  _updateIsActive(value: boolean) {
+    this.isActive = value;
+  }
 
   findWithOffset(options: any[], term: string, offset: number, skipDisabled = false): any {
     let typeAheadOptionMatcher = getOptionMatcher(this.args.typeAheadOptionMatcher || defaultTypeAheadMatcher, defaultTypeAheadMatcher, this.args.searchField);