+++ tests/dummy/config/targets.js
@@ -6,13 +6,14 @@ const browsers = [
   'last 1 Safari versions'
 ];
 
-const isCI = !!process.env.CI;
+const isCI = Boolean(process.env.CI);
 const isProduction = process.env.EMBER_ENV === 'production';
 
 if (isCI || isProduction) {
   browsers.push('ie 11');
 }
 
-module.exports = {
-  browsers
-};
+module.exports = [
+  ...browsers,
+  'maintained node versions'
+];