+++ tests/integration/components/power-select/opened-property-test.js
@@ -1,7 +1,7 @@
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
-import hbs from 'htmlbars-inline-precompile';
+import { hbs } from 'ember-cli-htmlbars';
 import { numbers } from '../constants';
 
 module('Integration | Component | Ember Power Select (The opened property)', function(hooks) {
@@ -13,6 +13,19 @@ module('Integration | Component | Ember Power Select (The opened property)', fun
     this.numbers = numbers;
     await render(hbs`
       <PowerSelect @options={{numbers}} @onChange={{action (mut foo)}} @initiallyOpened={{true}} as |option|>
+        {{option}}
+      </PowerSelect>
+    `);
+
+    assert.dom('.ember-power-select-dropdown').exists('Dropdown is opened');
+  });
+
+  test('[BUGFIX] the select can be rendered already opened by passing `@initiallyOpened={{true}}` AND `@selected`', async function(assert) {
+    assert.expect(1);
+
+    this.numbers = numbers;
+    await render(hbs`
+      <PowerSelect @options={{numbers}} @onChange={{action (mut foo)}} @initiallyOpened={{true}} @selected="seven" as |option|>
         {{option}}
       </PowerSelect>
     `);